+++ app/src/main/java/org/simple/clinic/navigation/v2/Router.kt
@@ -9,6 +9,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
 import org.simple.clinic.platform.analytics.Analytics
+import org.simple.clinic.util.setFragmentResult
 
 /**
  * Class that maintains a history of screens and is used to perform
@@ -413,14 +414,10 @@ class Router(
       currentTopScreen: NavRequest,
       screenResult: ScreenResult?
   ) {
-    val newTopNavRequest = history.top()
     if (currentTopScreen is ExpectingResult && screenResult != null) {
-      val targetFragment = fragmentManager.findFragmentByTag(newTopNavRequest.key.fragmentTag)
+      val requestType = currentTopScreen.requestType
 
-      require(targetFragment != null) { "Could not find fragment for key: [${newTopNavRequest.key}]" }
-      require(targetFragment is ExpectsResult) { "Key [${newTopNavRequest.key}] was pushed expecting results, but fragment [${targetFragment.javaClass.name}] does not implement [${ExpectsResult::class.java.name}]!" }
-
-      handler.post { targetFragment.onScreenResult(currentTopScreen.requestType, screenResult) }
+      fragmentManager.setFragmentResult(requestType, screenResult)
     }
   }